+++ CHANGELOG.md
@@ -24,6 +24,7 @@ Fixed:
 - Fixed incomplete autocannon block waterlogged state on completion
 - Fixed autocannon not syncing creative connections
 - Fixed incorrect top-bottom cannon cast merger fluid capacity
+- Fixed cannon builder assembly and layering collision
 
 ## [5.10.2] - 2026-01-17